Working csv2disk.frame; passes all tests in test-csv2disk.frame.
Revamped csv2disk.frame to better handle multiple files and chunking. Two basic methods for chunking: (1) split the csv files directly or (2) read in the csv files in chunks. Depending on the parameters chosen, (1) or (2) is attempted. The csv files are then saved as fst files in a specificed outdir, and a disk.frame is created at the end from the directory of fst files. This avoids repeatedly creating intermediate disk frames.

Also added ability to detect column classes and to estimate chunks based on csv file size and available memory.

Added test using nycflights13::flights. Fixed minor errors related to testing flights: POSIXct class actually returns as multiple classes, so that needed to be accounted for when estimating classes; and column classes did not need to be determined for bigreadr backend.

Commented out shard code, because disk.frame function is currently not setting metadata that the shard function expects to find.